Keep time entry method in sync between windows
I can imagine a use case where someone would not want this, and would prefer the old way where they could be out of sync. The problem is, participant_data was not designed with that in mind, so this makes consistent the way particiapnt_data expects the state of the room to look. I'm open to re-working participant_data and the time_entry key in order to support knowledge of different state across devices for one client.